Add RequestAwareBehavior and remove PerRequestBehavior
