--- src/Nut.MediatR.Behaviors/Validation/ValidationPerRequsetBehaviorBuilderExtensions.cs
@@ -1,14 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Nut.MediatR;
-
-public static class ValidationPerRequsetBehaviorBuilderExtensions
-{
-    public static PerRequsetBehaviorBuilder AddDataAnnotationValidation(this PerRequsetBehaviorBuilder builder)
-    {
-        builder.AddOpenBehavior(typeof(DataAnnotationValidationBehavior<,>));
-        return builder;
-    }
-}
